arch/arm/src/lpc17xx_40xx/lpc17_40_sdcard.c: Fix a type mismatch

+++ arch/arm/src/lpc17xx_40xx/lpc17_40_sdcard.c
@@ -382,7 +382,7 @@ static int  lpc17_40_sendcmd(FAR struct sdio_dev_s *dev, uint32_t cmd,
 static int  lpc17_40_recvsetup(FAR struct sdio_dev_s *dev,
               FAR uint8_t *buffer, size_t nbytes);
 static int  lpc17_40_sendsetup(FAR struct sdio_dev_s *dev,
-              FAR const uint8_t *buffer, uint32_t nbytes);
+              FAR const uint8_t *buffer, size_t nbytes);
 static int  lpc17_40_cancel(FAR struct sdio_dev_s *dev);
 
 static int  lpc17_40_waitresponse(FAR struct sdio_dev_s *dev, uint32_t cmd);
